Complaints - Added a functionality that gets the complaint region based on its lat and lng and stores it in the database

+++ Domain/ClientDTOs/Complaint/InsertComplaintAttachmentsDTO.cs
@@ -5,8 +5,8 @@ namespace Domain.ClientDTOs.Complaint
     public class InsertComplaintAttachmentsDTO
     {
         public IFormFile fileMedia { get; set; }
-        public float decLat { get; set; }
-        public float decLng { get; set; }
+        public decimal decLat { get; set; }
+        public decimal decLng { get; set; }
         public Boolean blnIsVideo { get; set; }
     }
 }
